feat: add cbor roundtrip test for Ballot
  Doing so, I had to reshuffle a bit the location of some of the generators; starting to move them more and more under the kernel. Ideally, we should contribute them up to Pallas.
