+++ Cargo.toml
@@ -15,14 +15,14 @@ default-members = ["crates/*"]
 resolver = "2"
 
 [workspace.dependencies]
+# External dependencies
 acto = { version = "0.7.4", features = ["tokio"] }
 anyhow = "1.0.95"
-async-trait = "0.1.83"
 async-compression = { version = "0.4.25",  features = ["tokio", "gzip"] }
+async-trait = "0.1.83"
 bech32 = "0.11.0"
 cbor4ii = { version = "1.0.0", features = ["serde1", "use_std"] }
 clap = { version = "4.5.38", features = ["derive", "env"] }
-reqwest = { version = "0.12.20", default-features = false, features = ["json", "stream", "rustls-tls"] }
 either = "1.15"
 futures-util = "0.3.31"
 gasket = { version = "0.8.0", features = ["derive"] } # More recent versions have regression (e.g. https://github.com/construkts/gasket-rs/pull/34)
@@ -32,6 +32,7 @@ indoc = "2.0"
 minicbor = "0.25.1" # Must match the version used by pallas
 num = { version = "0.4.3", default-features = false, features = ["alloc"] }
 opentelemetry = "0.30.0"
+reqwest = { version = "0.12.20", default-features = false, features = ["json", "stream", "rustls-tls"] }
 opentelemetry-otlp = { version = "0.30.0", features = [
     "grpc-tonic",
     "http-proto",
@@ -69,12 +70,15 @@ tracing-subscriber = { version = "0.3.18", features = [
 ] }
 typetag = "0.2.20"
 
+# Internal dependencies
+amaru = { path = "crates/amaru" }
 amaru-consensus = { path = "crates/amaru-consensus" }
 amaru-kernel = { path = "crates/amaru-kernel" }
 amaru-ledger = { path = "crates/amaru-ledger" }
 amaru-mempool = { path = "crates/amaru-mempool" }
 amaru-ouroboros = { path = "crates/ouroboros" }
 amaru-ouroboros-traits = { path = "crates/ouroboros-traits" }
+amaru-sim = { path = "crates/amaru-sim" }
 amaru-stores = { path = "crates/amaru-stores" }
 iter-borrow = { path = "crates/iter-borrow" }
 progress-bar = { path = "crates/progress-bar" }
@@ -90,13 +94,13 @@ kes-summed-ed25519 = { git = "https://github.com/txpipe/kes", rev = "f69fb357d46
 criterion = "0.6.0"
 ctor = "0.4.1"
 insta = "1.41.1"
+once_cell = "1.21.3"
 pretty_assertions = "1.4.1"
-proptest = { version = "1.5.0", default-features = false, features = ["alloc"] }
+proptest = { version = "1.5.0", default-features = false, features = ["std"] }
 rand = "0.9.1"
 rand_distr = "0.5.1"
 tempfile = "3.20.0"
 test-case = "3.3.1"
-once_cell = "1.21.3"
 
 # build-dependencies
 built = "0.8.0"
