+++ simulation/amaru-sim/Cargo.toml
@@ -11,6 +11,7 @@ documentation.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+# External dependencies ───────────────────────────────────────────────────────┐
 anyhow.workspace = true
 async-trait.workspace = true
 clap.workspace = true
@@ -20,9 +21,9 @@ hex.workspace = true
 pallas-crypto.workspace = true
 pallas-primitives.workspace = true
 parking_lot.workspace = true
-proptest = {workspace = true, features = ["std"] }
 pure-stage.workspace = true
 rand.workspace = true
+rand_distr.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "signal", "io-util"] }
@@ -31,16 +32,18 @@ tokio-util.workspace = true
 tracing-subscriber.workspace = true
 tracing.workspace = true
 
-amaru = { path = "../../crates/amaru" }
-amaru-kernel = { path = "../../crates/amaru-kernel" }
-amaru-consensus = { path = "../../crates/amaru-consensus" }
-amaru-ouroboros = { path = "../../crates/ouroboros" }
-amaru-stores = { path = "../../crates/amaru-stores" }
-slot-arithmetic = { path = "../../crates/slot-arithmetic" }
-rand_distr = "0.5.1"
+# Internal dependencies ───────────────────────────────────────────────────────┐
+amaru.workspace = true
+amaru-kernel.workspace = true
+amaru-consensus.workspace = true
+amaru-ouroboros.workspace = true
+amaru-stores.workspace = true
+slot-arithmetic.workspace = true
 
 [dev-dependencies]
+# External dependencies ───────────────────────────────────────────────────────┐
 insta = { workspace = true, features = ["json"] }
+proptest.workspace = true
 rand.workspace = true
 tempfile.workspace = true
 test-case.workspace = true
